requirements.txt
Django==2.1.15
django-mssql==1.8
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-python3-ldap==0.11.3
djangorestframework==3.10.3
